Web API - small refactor
