+++ TradingSystem/TradingSystem.WebApi/DTO/Store/StoreInfoActionDTO.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace TradingSystem.WebApi.DTO.Store.Permissions
+namespace TradingSystem.WebApi.DTO.Store
 {
     public class StoreInfoActionDTO
     {
         public string? Username { get; set; }
-        public Guid StoreID { get; set; }
+        public Guid StoreId { get; set; }
     }
 }
