--- TradingSystem/TradingSystem.WebApi/DTO/Store/StoreInfoDTO.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace TradingSystem.WebApi.DTO
-{
-    public class StoreInfoDTO
-    {
-        public Guid Id { get; set; }
-        public string? Name { get; set; }
-    }
-}
